+++ Tests/StreamerTests/Parser/Readium/ReadiumWebPubParserTests.swift
@@ -25,7 +25,7 @@ class ReadiumWebPubParserTests: XCTestCase {
         parser = ReadiumWebPubParser(pdfFactory: DefaultPDFDocumentFactory(), httpClient: DefaultHTTPClient())
 
         manifestAsset = FileAsset(file: fixtures.url(for: "flatland.json"))
-        manifestFetcher = FileFetcher(href: RelativeURL(path: "/flatland.json")!, file: manifestAsset.file)
+        manifestFetcher = FileFetcher(href: RelativeURL(path: "flatland.json")!, file: manifestAsset.file)
 
         packageAsset = FileAsset(file: fixtures.url(for: "audiotest.lcpa"))
         packageFetcher = try ArchiveFetcher(file: packageAsset.file)
@@ -46,33 +46,5 @@ class ReadiumWebPubParserTests: XCTestCase {
 
     func testAcceptsPackage() {
         XCTAssertNotNil(try parser.parse(asset: packageAsset, fetcher: packageFetcher, warnings: nil))
-    }
-
-    /// The `Link`s' hrefs are normalized to the `self` link for a manifest.
-    func testHrefsAreNormalizedToSelfForManifests() throws {
-        let publication = try XCTUnwrap(parser.parse(asset: manifestAsset, fetcher: manifestFetcher, warnings: nil)?.build())
-
-        XCTAssertEqual(
-            publication.readingOrder.map(\.href),
-            [
-                "http://www.archive.org/download/flatland_rg_librivox/flatland_1_abbott.mp3",
-                "https://readium.org/webpub-manifest/examples/Flatland/flatland_2_abbott.mp3",
-                "https://readium.org/flatland_3_abbott.mp3",
-            ]
-        )
-    }
-
-    /// The `Link`s' hrefs are normalized to `/` for a package.
-    func testHrefsAreNormalizedToRootForPackages() throws {
-        let publication = try XCTUnwrap(parser.parse(asset: packageAsset, fetcher: packageFetcher, warnings: nil)?.build())
-
-        XCTAssertEqual(
-            publication.readingOrder.map(\.href),
-            [
-                "http://readium.org/audio/gtr-jazz.mp3",
-                "/audio/Latin.mp3",
-                "/audio/oboe-bassoon.mp3",
-            ]
-        )
     }
 }